--- src/pages/PostItemPage.module.css
@@ -1,89 +1,0 @@
-.container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-  max-width: 45rem;
-  margin: 0 auto;
-  margin-top: 3rem;
-  padding: 0 1rem;
-}
-
-.inputWrapper {
-  width: 100%;
-  margin-bottom: 1.25rem;
-  flex-grow: 1;
-}
-
-.inputWrapper input {
-  width: 100%;
-}
-
-.label {
-  color: var(--color-gray-900);
-  font-weight: var(--font-weight-bold);
-  font-size: 1.5rem;
-  margin-bottom: 0.5rem;
-}
-
-.toggleButtonContainer {
-  width: 100%;
-  margin-bottom: 1.25rem;
-}
-
-.intro {
-  font-size: 1rem;
-  color: var(--color-gray-500);
-  font-size: 1rem;
-  margin-bottom: 1.3rem;
-}
-
-.colorOptions {
-  margin: 0 auto;
-  padding: 0;
-  display: flex;
-  flex-grow: 1;
-  gap: 1rem;
-  width: 100%;
-}
-
-.imageOptions {
-  display: flex;
-  flex-grow: 1;
-  gap: 1rem;
-  width: 100%;
-}
-
-.optionImage {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-  object-position: center;
-}
-
-.imageItem img {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-}
-
-.buttonContainer {
-  margin-top: 4rem;
-  width: 100%;
-}
-
-@media (max-width: 1024px) {
-  .colorOptions,
-  .imageOptions {
-    flex-wrap: wrap;
-  }
-
-  .inputWrapper {
-    width: 100%;
-  }
-
-  .inputWrapper input {
-    max-width: 100%;
-  }
-}
