--- src/pages/PostPage.jsx
@@ -1,5 +1,0 @@
-const PostPage = () => {
-  return <div>PostPage</div>;
-};
-
-export default PostPage;
